Fixed the error in the include BYTE_L2 and BYTE_R2, pushed all the headers to the controller.h.
I added the thread and mutex lib, and created a single thread only for updating the inputs. Also create the output function that will output all the keys to that was pressed.

Renamed check_sticks to update_inputs and created a function to be able to get the specific buttons and added almost all of the keys.

Removed the check_buttons